+++ LibCarla/source/carla/client/detail/Client.cpp
@@ -145,10 +145,6 @@ namespace detail {
     _pimpl->CallAndWait<void>("load_new_episode", std::move(map_name));
   }
 
-  bool Client::CheckIntermediateEpisode() {
-    return _pimpl->CallAndWait<bool>("check_intermediate_episode");
-  }
-
   void Client::CopyOpenDriveToServer(std::string opendrive, const rpc::OpendriveGenerationParameters & params) {
     // Await response, we need to be sure in this one.
     _pimpl->CallAndWait<void>("copy_opendrive_to_file", std::move(opendrive), params);
@@ -219,7 +215,7 @@ namespace detail {
   void Client::SetLightStateToVehicle(
       rpc::ActorId vehicle,
       const rpc::VehicleLightState &light_state) {
-    return _pimpl->AsyncCall("apply_vehicle_light_state", vehicle, light_state);
+    return _pimpl->AsyncCall("set_vehicle_light_state", vehicle, light_state);
   }
 
   rpc::Actor Client::SpawnActor(
@@ -233,6 +229,17 @@ namespace detail {
       const geom::Transform &transform,
       rpc::ActorId parent,
       rpc::AttachmentType attachment_type) {
+
+      if(attachment_type == rpc::AttachmentType::SpringArm) {
+        const auto a = transform.location.MakeSafeUnitVector(std::numeric_limits<float>::epsilon());
+        const auto z = geom::Vector3D(0.0f, 0.f, 1.0f);
+        constexpr float OneEps = 1.0f - std::numeric_limits<float>::epsilon();
+        if (geom::Math::Dot(a, z) > OneEps) {
+          std::cout << "WARNING: Transformations with translation only in the 'z' axis are ill-formed when \
+            using SprintArm attachment. Please, be careful with that." << std::endl;
+        }
+      }
+
     return _pimpl->CallAndWait<rpc::Actor>("spawn_actor_with_parent",
         description,
         transform,
@@ -315,13 +322,17 @@ namespace detail {
     _pimpl->AsyncCall("freeze_traffic_light", traffic_light, freeze);
   }
 
+  rpc::VehicleLightStateList Client::GetVehiclesLightStates() {
+    return _pimpl->CallAndWait<std::vector<std::pair<carla::ActorId, uint32_t>>>("get_vehicle_light_states");
+  }
+
   std::vector<ActorId> Client::GetGroupTrafficLights(rpc::ActorId traffic_light) {
     using return_t = std::vector<ActorId>;
     return _pimpl->CallAndWait<return_t>("get_group_traffic_lights", traffic_light);
   }
 
-  std::string Client::StartRecorder(std::string name) {
-    return _pimpl->CallAndWait<std::string>("start_recorder", name);
+  std::string Client::StartRecorder(std::string name, bool additional_data) {
+    return _pimpl->CallAndWait<std::string>("start_recorder", name, additional_data);
   }
 
   void Client::StopRecorder() {