+++ LibCarla/source/carla/client/detail/Simulator.cpp
@@ -90,7 +90,7 @@ namespace detail {
       using namespace std::literals::chrono_literals;
       _episode->WaitForState(10ms);
       auto episode = GetCurrentEpisode();
-      if (episode.GetId() != id && !_client.CheckIntermediateEpisode()) {
+      if (episode.GetId() != id) {
         return episode;
       }
     }
@@ -159,6 +159,10 @@ namespace detail {
   SharedPtr<BlueprintLibrary> Simulator::GetBlueprintLibrary() {
     auto defs = _client.GetActorDefinitions();
     return MakeShared<BlueprintLibrary>(std::move(defs));
+  }
+
+  rpc::VehicleLightStateList Simulator::GetVehiclesLightStates() {
+    return _client.GetVehiclesLightStates();
   }
 
   SharedPtr<Actor> Simulator::GetSpectator() {