+++ PythonAPI/carla/source/libcarla/Weather.cpp
@@ -13,11 +13,16 @@ namespace rpc {
 
   std::ostream &operator<<(std::ostream &out, const WeatherParameters &weather) {
     out << "WeatherParameters(cloudiness=" << std::to_string(weather.cloudiness)
+        << ", cloudiness=" << std::to_string(weather.cloudiness)
         << ", precipitation=" << std::to_string(weather.precipitation)
         << ", precipitation_deposits=" << std::to_string(weather.precipitation_deposits)
         << ", wind_intensity=" << std::to_string(weather.wind_intensity)
         << ", sun_azimuth_angle=" << std::to_string(weather.sun_azimuth_angle)
-        << ", sun_altitude_angle=" << std::to_string(weather.sun_altitude_angle) << ')';
+        << ", sun_altitude_angle=" << std::to_string(weather.sun_altitude_angle)
+        << ", fog_density=" << std::to_string(weather.fog_density)
+        << ", fog_distance=" << std::to_string(weather.fog_distance)
+        << ", fog_falloff=" << std::to_string(weather.fog_falloff)
+        << ", wetness=" << std::to_string(weather.wetness) << ')';
     return out;
   }
 
@@ -29,7 +34,7 @@ void export_weather() {
   namespace cr = carla::rpc;
 
   auto cls = class_<cr::WeatherParameters>("WeatherParameters")
-    .def(init<float, float, float, float, float, float, float, float, float>(
+    .def(init<float, float, float, float, float, float, float, float, float, float>(
         (arg("cloudiness")=0.0f,
          arg("precipitation")=0.0f,
          arg("precipitation_deposits")=0.0f,
@@ -38,6 +43,7 @@ void export_weather() {
          arg("sun_altitude_angle")=0.0f,
          arg("fog_density")=0.0f,
          arg("fog_distance")=0.0f,
+         arg("fog_falloff")=0.0f,
          arg("wetness")=0.0f)))
     .def_readwrite("cloudiness", &cr::WeatherParameters::cloudiness)
     .def_readwrite("precipitation", &cr::WeatherParameters::precipitation)
@@ -47,6 +53,7 @@ void export_weather() {
     .def_readwrite("sun_altitude_angle", &cr::WeatherParameters::sun_altitude_angle)
     .def_readwrite("fog_density", &cr::WeatherParameters::fog_density)
     .def_readwrite("fog_distance", &cr::WeatherParameters::fog_distance)
+    .def_readwrite("fog_falloff", &cr::WeatherParameters::fog_falloff)
     .def_readwrite("wetness", &cr::WeatherParameters::wetness)
     .def("__eq__", &cr::WeatherParameters::operator==)
     .def("__ne__", &cr::WeatherParameters::operator!=)