+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Lights/CarlaLight.cpp
@@ -6,7 +6,7 @@
 
 #include "CarlaLight.h"
 #include "CarlaLightSubsystem.h"
-
+#include "Carla/Game/CarlaStatics.h"
 
 UCarlaLight::UCarlaLight()
 {
@@ -46,7 +46,7 @@ void UCarlaLight::SetLightIntensity(float Intensity)
   UpdateLights();
 }
 
-float UCarlaLight::GetLightIntensity()
+float UCarlaLight::GetLightIntensity() const
 {
   return LightIntensity;
 }
@@ -55,9 +55,10 @@ void UCarlaLight::SetLightColor(FLinearColor Color)
 {
   LightColor = Color;
   UpdateLights();
+  RecordLightChange();
 }
 
-FLinearColor UCarlaLight::GetLightColor()
+FLinearColor UCarlaLight::GetLightColor() const
 {
   return LightColor;
 }
@@ -66,9 +67,10 @@ void UCarlaLight::SetLightOn(bool bOn)
 {
   bLightOn = bOn;
   UpdateLights();
+  RecordLightChange();
 }
 
-bool UCarlaLight::GetLightOn()
+bool UCarlaLight::GetLightOn() const
 {
   return bLightOn;
 }
@@ -78,7 +80,7 @@ void UCarlaLight::SetLightType(ELightType Type)
   LightType = Type;
 }
 
-ELightType UCarlaLight::GetLightType()
+ELightType UCarlaLight::GetLightType() const
 {
   return LightType;
 }
@@ -105,6 +107,7 @@ void UCarlaLight::SetLightState(carla::rpc::LightState LightState)
   LightType = static_cast<ELightType>(LightState._group);
   bLightOn = LightState._active;
   UpdateLights();
+  RecordLightChange();
 }
 
 FVector UCarlaLight::GetLocation() const
@@ -112,7 +115,25 @@ FVector UCarlaLight::GetLocation() const
   return GetOwner()->GetActorLocation();
 }
 
-uint32 UCarlaLight::GetId() const
+int UCarlaLight::GetId() const
 {
-  return GetUniqueID();
+  return Id;
+}
+
+void UCarlaLight::SetId(int InId)
+{
+  Id = InId;
+}
+
+void UCarlaLight::RecordLightChange() const
+{
+  auto* Episode = UCarlaStatics::GetCurrentEpisode(GetWorld());
+  if (Episode)
+  {
+    auto* Recorder = Episode->GetRecorder();
+    if (Recorder && Recorder->IsEnabled())
+    {
+      Recorder->AddEventLightSceneChanged(this);
+    }
+  }
 }