+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Recorder/CarlaRecorder.h
@@ -11,6 +11,11 @@
 
 #include "Carla/Actor/ActorDescription.h"
 
+#include "CarlaRecorderPlatformTime.h"
+#include "CarlaRecorderBoundingBox.h"
+#include "CarlaRecorderKinematics.h"
+#include "CarlaRecorderLightScene.h"
+#include "CarlaRecorderLightVehicle.h"
 #include "CarlaRecorderAnimVehicle.h"
 #include "CarlaRecorderAnimWalker.h"
 #include "CarlaRecorderCollision.h"
@@ -28,6 +33,8 @@
 
 class AActor;
 class UCarlaEpisode;
+class ACarlaWheeledVehicle;
+class UCarlaLight;
 
 enum class CarlaRecorderPacketId : uint8_t
 {
@@ -40,7 +47,12 @@ enum class CarlaRecorderPacketId : uint8_t
   Position,
   State,
   AnimVehicle,
-  AnimWalker
+  AnimWalker,
+  VehicleLight,
+  SceneLight,
+  Kinematics,
+  BoundingBox,
+  PlatformTime
 };
 
 /// Recorder for the simulation
@@ -64,7 +76,7 @@ public:
   void Disable(void);
 
   // start / stop
-  std::string Start(std::string Name, FString MapName);
+  std::string Start(std::string Name, FString MapName, bool AdditionalData = false);
 
   void Stop(void);
 
@@ -88,6 +100,14 @@ public:
   void AddAnimVehicle(const CarlaRecorderAnimVehicle &Vehicle);
 
   void AddAnimWalker(const CarlaRecorderAnimWalker &Walker);
+
+  void AddLightVehicle(const CarlaRecorderLightVehicle &LightVehicle);
+
+  void AddEventLightSceneChanged(const UCarlaLight* Light);
+
+  void AddKinematics(const CarlaRecorderKinematics &ActorKinematics);
+
+  void AddBoundingBox(const CarlaRecorderBoundingBox &ActorBoundingBox);
 
   // set episode
   void SetEpisode(UCarlaEpisode *ThisEpisode)
@@ -124,6 +144,9 @@ private:
 
   bool Enabled;   // enabled or not
 
+  // enabling this records additional data (kinematics, bounding boxes, etc)
+  bool bAdditionalData = false;
+
   uint32_t NextCollisionId = 0;
 
   // files
@@ -142,6 +165,12 @@ private:
   CarlaRecorderStates States;
   CarlaRecorderAnimVehicles Vehicles;
   CarlaRecorderAnimWalkers Walkers;
+  CarlaRecorderLightVehicles LightVehicles;
+  CarlaRecorderLightScenes LightScenes;
+  CarlaRecorderActorsKinematics Kinematics;
+  CarlaRecorderBoundingBoxes BoundingBoxes;
+  CarlaRecorderPlatformTime PlatformTime;
+
 
   // replayer
   CarlaReplayer Replayer;
@@ -154,4 +183,7 @@ private:
   void AddWalkerAnimation(FActorView &View);
   void AddVehicleAnimation(FActorView &View);
   void AddTrafficLightState(FActorView &View);
+  void AddVehicleLight(FActorView &View);
+  void AddActorKinematics(FActorView &View);
+  void AddActorBoundingBox(FActorView &View);
 };