+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Server/CarlaServer.cpp
@@ -35,6 +35,7 @@
 #include <carla/rpc/VehicleControl.h>
 #include <carla/rpc/VehiclePhysicsControl.h>
 #include <carla/rpc/VehicleLightState.h>
+#include <carla/rpc/VehicleLightStateList.h>
 #include <carla/rpc/WalkerBoneControl.h>
 #include <carla/rpc/WalkerControl.h>
 #include <carla/rpc/WeatherParameters.h>
@@ -68,7 +69,7 @@ public:
   FPimpl(uint16_t RPCPort, uint16_t StreamingPort)
     : Server(RPCPort),
       StreamingServer(StreamingPort),
-      BroadcastStream(StreamingServer.MakeMultiStream())
+      BroadcastStream(StreamingServer.MakeStream())
   {
     BindActions();
   }
@@ -80,7 +81,7 @@ public:
 
   carla::streaming::Server StreamingServer;
 
-  carla::streaming::MultiStream BroadcastStream;
+  carla::streaming::Stream BroadcastStream;
 
   UCarlaEpisode *Episode = nullptr;
 
@@ -228,35 +229,14 @@ void FCarlaServer::FPimpl::BindActions()
   BIND_SYNC(load_new_episode) << [this](const std::string &map_name) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
-    FString MapName = cr::ToFString(map_name);
-    MapName = MapName.IsEmpty() ? Episode->GetMapName() : MapName;
-    auto Maps = UCarlaStatics::GetAllMapNames();
-    Maps.Add("OpenDriveMap");
-    bool bMissingMap = true;
-    for (auto & Map : Maps)
-    {
-      if(Map.Contains(MapName))
-      {
-        bMissingMap = false;
-        break;
-      }
-    }
-    if(bMissingMap)
+    if(!Episode->LoadNewEpisode(cr::ToFString(map_name)))
     {
       RESPOND_ERROR("map not found");
     }
-    UCarlaStatics::GetGameInstance(Episode->GetWorld())->SetMapToLoad(MapName);
-    Episode->LoadNewEpisode(cr::ToFString("EmptyMap"));
     return R<void>::Success();
   };
 
-  BIND_SYNC(check_intermediate_episode) << [this]() -> R<bool>
-  {
-    REQUIRE_CARLA_EPISODE();
-    return UCarlaStatics::GetGameInstance(Episode->GetWorld())->IsLevelPendingLoad();
-  };
-
-  BIND_SYNC(copy_opendrive_to_file) << [this](const std::string &opendrive, carla::rpc::OpendriveGenerationParameters Params) -> R<void>
+  BIND_SYNC(copy_opendrive_to_file) << [this](const std::string &opendrive, cr::OpendriveGenerationParameters Params) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     if (!Episode->LoadNewOpendriveEpisode(cr::ToLongFString(opendrive), Params))
@@ -665,7 +645,7 @@ void FCarlaServer::FPimpl::BindActions()
     return R<void>::Success();
   };
 
-  BIND_SYNC(apply_vehicle_light_state) << [this](
+  BIND_SYNC(set_vehicle_light_state) << [this](
       cr::ActorId ActorId,
       cr::VehicleLightState LightState) -> R<void>
   {
@@ -895,6 +875,27 @@ void FCarlaServer::FPimpl::BindActions()
     return R<void>::Success();
   };
 
+  BIND_SYNC(get_vehicle_light_states) << [this]() -> R<cr::VehicleLightStateList>
+  {
+    REQUIRE_CARLA_EPISODE();
+    cr::VehicleLightStateList List;
+
+    auto It = Episode->GetActorRegistry().begin();
+    for (; It != Episode->GetActorRegistry().end(); ++It)
+    {
+      auto Actor = It->GetActor();
+      if (!Actor->IsPendingKill() && It->GetActorType() == FActorView::ActorType::Vehicle)
+      {
+        const ACarlaWheeledVehicle *Vehicle = Cast<ACarlaWheeledVehicle>(Actor);
+        List.emplace_back(
+            It->GetActorId(),
+            cr::VehicleLightState(Vehicle->GetVehicleLightState()).GetLightStateAsValue());
+      }
+    }
+
+    return List;
+  };
+
   BIND_SYNC(get_group_traffic_lights) << [this](
       const cr::ActorId ActorId) -> R<std::vector<cr::ActorId>>
   {
@@ -923,10 +924,10 @@ void FCarlaServer::FPimpl::BindActions()
 
   // ~~ Logging and playback ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-  BIND_SYNC(start_recorder) << [this](std::string name) -> R<std::string>
+  BIND_SYNC(start_recorder) << [this](std::string name, bool AdditionalData) -> R<std::string>
   {
     REQUIRE_CARLA_EPISODE();
-    return R<std::string>(Episode->StartRecorder(name));
+    return R<std::string>(Episode->StartRecorder(name, AdditionalData));
   };
 
   BIND_SYNC(stop_recorder) << [this]() -> R<void>
@@ -1057,6 +1058,7 @@ void FCarlaServer::FPimpl::BindActions()
       [=](auto, const C::SetSimulatePhysics &c) {   MAKE_RESULT(set_actor_simulate_physics(c.actor, c.enabled)); },
       // TODO: SetAutopilot should be removed. This is the old way to control the vehicles
       [=](auto, const C::SetAutopilot &c) {         MAKE_RESULT(set_actor_autopilot(c.actor, c.enabled)); },
+      [=](auto, const C::SetVehicleLightState &c) { MAKE_RESULT(set_vehicle_light_state(c.actor, c.light_state)); },
       [=](auto, const C::ApplyWalkerState &c) {     MAKE_RESULT(set_walker_state(c.actor, c.transform, c.speed)); });
 
 #undef MAKE_RESULT