+++ CHANGELOG.md
@@ -1,18 +1,35 @@
 ## Latest
 
+  * Added Light ids
+  * Added vehicle light and street light data to recorder
   * Added API function `add_angular_impulse()` to add angular impulse to any actor
   * Fixed rain drop spawn issues when spawning camera sensors
   * Fixed Update.sh from failing when the root folder contains a space on it
   * Fixed colors of lane markings when importing a map, they were reversed (white and yellow)
   * Fixed missing include directive in file **WheelPhysicsControl.h**
   * Fixed gravity measurement bug from IMU sensor
+  * All sensors are now multi-stream, that means that the same sensor can be listened from different clients
+  * Fixed point cloud of LiDAR. Now the points are given correctly in the sensor's coordinate system.
+  * Exposed matrix form of transformation to the client and Python API.
+  * Added PythonAPI command to set multiple car light states at once
+  * Added PythonAPI `carla.world.get_vehicles_light_states` to get all the car light states at once
   * OpenDRIVE ingestion bugfixes
+  * Added a warning if the user tries to use the SpringArm exactly in the 'z' axis of the attached actor
+  * Improved the LiDAR and Radar sensors with a parallel implentation of the raycasting
+  * Added an approximation of the intensity of each point of the cloud in the LiDAR sensor
   * Added Dynamic Vision Sensor (DVS) camera based on ESIM simulation http://rpg.ifi.uzh.ch/esim.html
+  * Improved LiDAR and radar to better match the shape of the vehicles
+  * Added support for additional TraCI clients in Sumo co-simulation
   * Added API functions `get_right_vector` and `get_up_vector`
+  * Added default values and a warning message for lanes missing the width parameter in OpenDRIVE
   * Added parameter to enable/disable pedestrian navigation in standalone mode
   * Improved mesh split in standalone mode
+  * Fixed delay in the tcp communication from server to client, improving performance in synchronous mode in linux systems
   * Fixed large RAM usage when loading polinomial geometry from OpenDRIVE
-
+  * Fixed collision issues when debug draw(debug.draw_line) is called
+  * Fixed Gyroscope sensor to properly give angular velocity readings in local frame
+  * Added Renderdoc plugin to the Unreal project
+  * Replace deprectated `platform.dist()` with recommended `distro.linux_distribution()`
 
 ## CARLA 0.9.9